some changes

+++ GraviSim/src/Object.h
@@ -14,7 +14,7 @@ private:
 	class Properties {
 	public:
 		float mass;					 //in 10^24 Kg
-		int radius;					 //in px --- 1 px = 10^3 km
+		float radius;				 //in px --- 1 px = 10^3 km
 		sf::Vector2f pos;			 //in km --- 1 px = 10^3 km
 		sf::Vector2f speed;			 //in km/s
 		sf::Vector2f acceleration;	 //in m/s^2
@@ -41,7 +41,7 @@ public:
 	bool useForGravityCenter;
 	Properties properties;
 
-	Object(float mass, int radius, sf::Vector2f pos, sf::Vector2f speed = sf::Vector2f(0, 0), std::string name = "", sf::Color color = sf::Color(255, 255, 255), bool fixed = false);
+	Object(float mass, float radius, sf::Vector2f pos, sf::Vector2f speed = sf::Vector2f(0, 0), std::string name = "", sf::Color color = sf::Color(255, 255, 255), bool fixed = false);
 	Object(Properties properties);
 	Object();
